add connected wrapper for the small grid item that owns upload status, swap into the new view
